requirements.txt
pyperclip
Pillow
requests
